feat-api(dados proposta analise): retornando classificacao_score

+++ database/migrations/2021_05_31_175433_create_cad_analise_pessoa_proposta.php
@@ -45,10 +45,10 @@ class CreateCadAnalisePessoaProposta extends Migration
                 ->nullable()
                 ->unsigned();
 
-            $table->integer('restricao')
+            $table->float('restricao', 8, 2)
                 ->nullable();
 
-            $table->integer('score')
+            $table->float('score', 8, 2)
                 ->nullable();
 
             $table->string('classificacao_score')
